Give resources unique names to avoid resource cloning
When this cookbook is used to install multiple
versions of nodejs on a single machine, the provider
creates multiple resources for `directory`, `git`, and
`template`, resulting in a clone.

Resource cloning is deprecated and an error in Chef 13.

https://docs.chef.io/deprecations_resource_cloning.html

This change is almost certainly not the right way to do it, but
works.